BAU - reorganise test classes
- Reorganise test classes (including integration tests) to mirror packages
- Updated few integration tests suffix to IT so fail safe plugin can run them as integration tests
